feat(auth-refactor): move wallet upgrade and sync to wallet folder
