+++ packages/blockchain-wallet-v4-frontend/src/data/auth/sagas.ts
@@ -33,41 +33,6 @@ export default ({ api, coreSagas, networks }) => {
     networks
   })
 
-  const forceSyncWallet = function* () {
-    yield put(actions.core.walletSync.forceSync())
-    const { error } = yield race({
-      error: take(actionTypes.core.walletSync.SYNC_ERROR),
-      success: take(actionTypes.core.walletSync.SYNC_SUCCESS)
-    })
-    if (error) {
-      throw new Error('Sync failed')
-    }
-  }
-
-  const upgradeWallet = function* (action) {
-    try {
-      const { version } = action.payload
-      const password = yield call(promptForSecondPassword)
-      // eslint-disable-next-line default-case
-      switch (version) {
-        case 3:
-          yield coreSagas.wallet.upgradeToV3({ password })
-          break
-        case 4:
-          yield coreSagas.wallet.upgradeToV4({ password })
-          break
-        default:
-          break
-      }
-      yield call(forceSyncWallet)
-    } catch (e) {
-      if (e.message === 'Already a v4 wallet') return
-      yield put(actions.logs.logErrorMessage(logLocation, 'upgradeWallet', e))
-      yield put(actions.alerts.displayError(C.WALLET_UPGRADE_ERROR))
-      yield put(actions.modals.closeModal())
-    }
-  }
-
   const upgradeAddressLabelsSaga = function* () {
     const addressLabelSize = yield call(coreSagas.kvStore.btc.fetchMetadataBtc)
     if (addressLabelSize > 100) {
@@ -218,7 +183,7 @@ export default ({ api, coreSagas, networks }) => {
       // If needed, the user should upgrade its wallet before being able to open the wallet
       const isHdWallet = yield select(selectors.core.wallet.isHdWallet)
       if (!isHdWallet) {
-        yield put(actions.auth.upgradeWallet(3))
+        yield put(actions.wallet.upgradeWallet(3))
         yield take(actionTypes.core.walletSync.SYNC_SUCCESS)
       }
       const isLatestVersion = yield select(selectors.core.wallet.isWrapperLatestVersion)
@@ -228,7 +193,7 @@ export default ({ api, coreSagas, networks }) => {
         .getOrElse(DEFAULT_INVITATIONS)
       const isSegwitEnabled = invitations.segwit
       if (!isLatestVersion && isSegwitEnabled) {
-        yield put(actions.auth.upgradeWallet(4))
+        yield put(actions.wallet.upgradeWallet(4))
         yield take(actionTypes.core.walletSync.SYNC_SUCCESS)
       }
       const isAccountReset: boolean = yield select(selectors.auth.getAccountReset)
@@ -830,7 +795,6 @@ export default ({ api, coreSagas, networks }) => {
     saveGoals,
     startSockets,
     triggerWalletMagicLink,
-    upgradeAddressLabelsSaga,
-    upgradeWallet
+    upgradeAddressLabelsSaga
   }
 }
