+++ packages/blockchain-wallet-v4-frontend/src/data/auth/slice.ts
@@ -148,8 +148,7 @@ const authSlice = createSlice({
     triggerWalletMagicLinkFailure: () => {},
     triggerWalletMagicLinkLoading: () => {},
     triggerWalletMagicLinkNotAsked: () => {},
-    triggerWalletMagicLinkSuccess: () => {},
-    upgradeWallet: (state, action) => {}
+    triggerWalletMagicLinkSuccess: () => {}
   }
 })
