+++ packages/blockchain-wallet-v4-frontend/src/data/wallet/actionTypes.ts
@@ -19,3 +19,6 @@ export const SUBMIT_CONFIRMATION = '@WALLET.SUBMIT_CONFIRMATION'
 
 // MANAGE
 export const MANAGE_WALLET_SELECTION = '@WALLET.MANAGE_OPTIONS'
+
+// UPGRADE
+export const UPGRADE_WALLET = '@WALLET.UPGRADE_WALLET'
