+++ packages/blockchain-wallet-v4-frontend/src/data/wallet/sagaRegister.ts
@@ -12,5 +12,6 @@ export default ({ coreSagas }) => {
     yield takeEvery(AT.VERIFY_MNEMONIC, walletSagas.verifyMnemonic)
     yield takeEvery(AT.EDIT_BTC_ACCOUNT_LABEL, walletSagas.editBtcAccountLabel)
     yield takeLatest(AT.SET_MAIN_PASSWORD, walletSagas.setMainPassword)
+    yield takeLatest(AT.UPGRADE_WALLET, walletSagas.upgradeWallet)
   }
 }
